+++ pkg/collector/corechecks/snmp/report_metrics.go
@@ -7,6 +7,8 @@ import (
 	"github.com/DataDog/datadog-agent/pkg/aggregator"
 	"github.com/DataDog/datadog-agent/pkg/metrics"
 	"github.com/DataDog/datadog-agent/pkg/util/log"
+
+	"github.com/DataDog/datadog-agent/pkg/collector/corechecks/snmp/common"
 )
 
 type metricSender struct {
@@ -50,7 +52,7 @@ func (ms *metricSender) reportScalarMetrics(metric MetricsConfig, values *result
 		return
 	}
 
-	scalarTags := copyStrings(tags)
+	scalarTags := common.CopyStrings(tags)
 	scalarTags = append(scalarTags, metric.getSymbolTags()...)
 	ms.sendMetric(metric.Symbol.Name, value, scalarTags, metric.ForcedType, metric.Options, metric.Symbol.extractValuePattern)
 }
@@ -66,7 +68,7 @@ func (ms *metricSender) reportColumnMetrics(metricConfig MetricsConfig, values *
 		for fullIndex, value := range metricValues {
 			// cache row tags by fullIndex to avoid rebuilding it for every column rows
 			if _, ok := rowTagsCache[fullIndex]; !ok {
-				rowTagsCache[fullIndex] = append(copyStrings(tags), metricConfig.getTags(fullIndex, values)...)
+				rowTagsCache[fullIndex] = append(common.CopyStrings(tags), metricConfig.getTags(fullIndex, values)...)
 				log.Debugf("report column: caching tags `%v` for fullIndex `%s`", rowTagsCache[fullIndex], fullIndex)
 			}
 			rowTags := rowTagsCache[fullIndex]
@@ -140,22 +142,22 @@ func (ms *metricSender) sendMetric(metricName string, value snmpValueType, tags 
 
 func (ms *metricSender) gauge(metric string, value float64, hostname string, tags []string) {
 	// we need copy tags before using sender due to https://github.com/DataDog/datadog-agent/issues/7159
-	ms.sender.Gauge(metric, value, hostname, copyStrings(tags))
+	ms.sender.Gauge(metric, value, hostname, common.CopyStrings(tags))
 }
 
 func (ms *metricSender) rate(metric string, value float64, hostname string, tags []string) {
 	// we need copy tags before using sender due to https://github.com/DataDog/datadog-agent/issues/7159
-	ms.sender.Rate(metric, value, hostname, copyStrings(tags))
+	ms.sender.Rate(metric, value, hostname, common.CopyStrings(tags))
 }
 
 func (ms *metricSender) monotonicCount(metric string, value float64, hostname string, tags []string) {
 	// we need copy tags before using sender due to https://github.com/DataDog/datadog-agent/issues/7159
-	ms.sender.MonotonicCount(metric, value, hostname, copyStrings(tags))
+	ms.sender.MonotonicCount(metric, value, hostname, common.CopyStrings(tags))
 }
 
 func (ms *metricSender) serviceCheck(checkName string, status metrics.ServiceCheckStatus, hostname string, tags []string, message string) {
 	// we need copy tags before using sender due to https://github.com/DataDog/datadog-agent/issues/7159
-	ms.sender.ServiceCheck(checkName, status, hostname, copyStrings(tags), message)
+	ms.sender.ServiceCheck(checkName, status, hostname, common.CopyStrings(tags), message)
 }
 
 func getFlagStreamValue(placement uint, strValue string) (float64, error) {
